backend/requirements.txt
fastapi
uvicorn
pandas
numpy

# api_key
python-dotenv
matplotlib
plotly
scipy
scikit-learn
pinecone-client==2.2.2
bs4
beautifulsoup4
tiktoken
nltk
webdriver_manager==4.0.1
selenium==4.13.0
langchain==0.1.8
langchain-community==0.0.21
langchain-core==0.1.24
langchain-openai==0.0.6
langchainplus-sdk==0.0.20
openai==1.12.0